backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-q2==1.8.0
psycopg[binary]==3.3.3
python-dotenv==1.1.0
dj-database-url==2.3.0
whitenoise==6.9.0
gunicorn==23.0.0

pytest==8.3.5
pytest-django==4.11.1
django-cors-headers==3.13.0
whitenoise